3. Интерфейс Input [#181092 #209569]

+++ src/main/java/ru/job4j/poly/Transpor.java
@@ -5,5 +5,5 @@ public interface Transpor {
 
      void passengers(int numbers);
 
-     int refuelPrice(double liters);
+     double refuelPrice(double liters);
 }
